prepare basic structure for integration tests with kind-cluster sandbox.
