+++ integration/immutable_data/resolve/plain-yaml-json/_namespace.json
@@ -2,6 +2,6 @@
   "apiVersion": "v1",
   "kind": "Namespace",
   "metadata": {
-    "name": "kubectl-envsubst-integration-tests-ns-2"
+    "name": "kubectl-envsubst-integration-tests-ns-1"
   }
 }
