+++ integration/immutable_data/resolve/plain-yaml-json/secret.yml
@@ -3,7 +3,7 @@ apiVersion: v1
 kind: Secret
 metadata:
   name: my-app
-  namespace: kubectl-envsubst-integration-tests-ns-2
+  namespace: kubectl-envsubst-integration-tests-ns-1
 type: Opaque
 data:
   username: bXktYXBwLXVzZXI=
