+++ integration/immutable_data/resolve/plain-yaml-json/service.yml
@@ -3,7 +3,7 @@ apiVersion: v1
 kind: Service
 metadata:
   name: my-app
-  namespace: kubectl-envsubst-integration-tests-ns-2
+  namespace: kubectl-envsubst-integration-tests-ns-1
 spec:
   selector:
     app: my-app
