+++ integration/immutable_data/resolve/plain-yaml-json/serviceaccount.json
@@ -3,6 +3,6 @@
   "kind": "ServiceAccount",
   "metadata": {
     "name": "my-app",
-    "namespace": "kubectl-envsubst-integration-tests-ns-2"
+    "namespace": "kubectl-envsubst-integration-tests-ns-1"
   }
 }
